adding kotlin-couroutines for progressbar implementation, mostly done the features

+++ app/src/main/java/com/example/githubuserdetailed/api/Envelope.kt
@@ -1,3 +1,3 @@
 package com.example.githubuserdetailed.api
 
-data class Envelope<T>(var total_count: Double, var incomplete_results: Boolean, var items: T)
+data class Envelope<T>(var total_count: Int, var incomplete_results: Boolean, var items: T)
